+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Issue #521 - Apply Default API settings when creating an api query
  - Issue #532 - Code Action "Add multiple fields" doesn't work when SourceTable includes a dot
  - Issue #534 - Text Decorators does not work correctly when app.json is saved as UTF8-BOM
+ - Issue #535 - Separate 'Sort Triggers' action #535
  - Issue #536 - [TEST] is not converted to [Test] when running "fix Case"
  - Issue #538 - Issues running cleanup: "Unexpected StringComparison type: 'CurrentCultureIgnoreCase'. Only 'Ordinal' and 'OrdinalIgnoreCase' are supported."
 
@@ -12,8 +13,9 @@ Thank you
  - jorispoppe for reporting issue #521
  - achim-t for reporting issue #532
  - zabcik for reporting issue #534
+ - fvet for reporting issue #535
  - guidorobben for reporting issue #356
- - jwikman for reporting issue #538
+ - jwikman for reporting issue #538 
 
 ## 3.0.56
  - Issue #507 - Namespaces quick fixes and commands